requirements.txt
python-dateutil
zimscraperlib
requests>=2.23,<3.0
Pillow==7.1.2
beautifulsoup4==4.9.0
